### 2019-02-19 #1263 @whedon
[PRE REVIEW]: riskclustr: Functions to Study Etiologic Heterogeneity
**Submitting author:** @zabore (<a href="http://orcid.org/0000-0002-1402-4498">Emily Zabor</a>)
**Repository:** <a href="https://github.com/zabore/riskclustr" target ="_blank">https://github.com/zabore/riskclustr</a>
**Version:** 0.1.0
