### 2019-02-19 #1263 @pjotrp
Very cool paper :)
