### 2019-02-20 #1269 @pjotrp
@avhoeck is our reviewer!

@zabore, we are starting review in this issue tracker. To expedite the review process do you mind going through above list of check boxes and make sure they can be ticked (you can't tick them). Also check the PDF output carefully. Ping us here when you are done.
