### 2019-03-18 #1269 @zabore
I am so sorry that I overlooked the initial email requesting me to go over
that checklist. I hadn't realized that anything required my attention. I
just went through the list, and made one change to the package website to
address contributions to the package, everything else was already
satisfied. So I believe now that everything in the list can be checked.

Thanks, and sorry again for this oversight.
Emily

On Sun, Mar 17, 2019 at 10:00 AM Pjotr Prins <notifications@github.com>
wrote:

> Actually we are waiting for the author to confirm we can tick all boxes.
> @zabore <https://github.com/zabore> Emily, do you mind responding?
>
