### 2019-03-18 #1269 @avhoeck
Great!

________________________________
Van: Pjotr Prins [notifications@github.com]
Verzonden: maandag 18 maart 2019 16:05
Aan: openjournals/joss-reviews
CC: Hoeck, A. van; Mention
Onderwerp: Re: [openjournals/joss-reviews] [REVIEW]: riskclustr: Functions to Study Etiologic Heterogeneity (#1269)

Thank you @zabore<https://github.com/zabore>. @avhoeck<https://github.com/avhoeck> you can start review.
