+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: riskclustr
 Type: Package
 Title: Functions to Study Etiologic Heterogeneity
-Version: 0.1.0
+Version: 0.1.1
 Authors@R: person(given = "Emily C.", family = "Zabor", email = "zabore@mskcc.org", role = c("aut", "cre"))
 Description: A collection of functions related to the study of etiologic heterogeneity both across disease subtypes and across individual disease markers. The included functions allow one to quantify the extent of etiologic heterogeneity in the context of a case-control study, and provide p-values to test for etiologic heterogeneity across individual risk factors.
 Depends:
