### 2019-03-20 2050576
fixing a bug when only 1 risk factor is passed
