### 2019-03-26 #1269 @avhoeck
This is a nice package that builds on functionalities from previous work developed by the publishing author. Overall, this submission fulfills all requirements as stated by https://joss.readthedocs.io/en/latest/review_criteria.html and I'm ready to sign off as it is.

Some minor changes that could improve the submission:

Previous work focusing on etiologic heterogeneity that has been published by E. Zabor (doi: 10.1002/sim.7405 and 10.1002/sim.5902) was together with C. Begg. If C. Begg has no contributions in this work, the authorship button can be signed off.

On http://www.emilyzabor.com/riskclustr/ website under “reference”, I would expect to see the manuscript DOI or URL, and not an overview of all the riskcluser core functions. Moreover, it would be nice to include (either under “reference” or on the main page) a link to the riskclustr manuscript anyway which is a good resource for the user.

The source code is explicit including bug reporting and extra clarifications for basically every step. I have successfully installed the package and no errors appeared upon testing. However, the simulated data is a bit too basic for me. Would it be possible to rename the column names of simulated data frame (eg X1… into dummy risk factors and Y1-Y300 into dummy sample/patient names?). Moreover, I dont see whether Y1-Y300 scores are actually taken into account in any of the functionalities. If not, I would remove these from the simulated dataframe, an otherwise, I would include an extra argument that refers to observed sample scores for the functions that need sample observations (I guess all?). Ideally, would it be feasible to include Cancer and Steroid Hormone (CASH) study and the Womens’ Contraceptive and Reproductive Experiences (CARE) real you analysed previously to analyse “real-world” datasets.

Thanks for your work!
All the best,
Arne
