### 2019-03-26 #1269 @zabore
Arne,
Thank you for these comments.
C. Begg did not make any contribution to the development of this software
package, though you are correct that he contributed to the original
methodology.
I appreciate your comment about the website header "Reference" and will
change this, and link elsewhere to the methodology paper, likely in the
introduction/overview.
With respect to the simulated data, the continuous tumor markers, currently
labeled y1-y30, are used to demonstrate functionality of the
"optimal_kmeans_d" function in "Tutorial: Identify the optimal D subtype
solution". If you think it would help, I could rename the x variables to
start with "rf" for risk factor and the y variables to start with "tm" for
tumor marker if this would increase clarity, but it's true that these
simulated data are quite generic regardless and for demonstration purposes
only. Unfortunately neither the CASH nor CARE data can be made public at
this time.

Thanks again,
Emily

On Tue, Mar 26, 2019 at 6:05 AM Arne Van Hoeck <notifications@github.com>
wrote:

> This is a nice package that builds on functionalities from previous work
> developed by the publishing author. Overall, this submission fulfills all
> requirements as stated by
> https://joss.readthedocs.io/en/latest/review_criteria.html and I'm ready
> to sign off as it is.
>
> Some minor changes that could improve the submission:
>
> Previous work focusing on etiologic heterogeneity that has been published
> by E. Zabor (doi: 10.1002/sim.7405 and 10.1002/sim.5902) was together with
> C. Begg. If C. Begg has no contributions in this work, the authorship
> button can be signed off.
>
> On http://www.emilyzabor.com/riskclustr/ website under “reference”, I
> would expect to see the manuscript DOI or URL, and not an overview of all
> the riskcluser core functions. Moreover, it would be nice to include
> (either under “reference” or on the main page) a link to the riskclustr
> manuscript anyway which is a good resource for the user.
>
> The source code is explicit including bug reporting and extra
> clarifications for basically every step. I have successfully installed the
> package and no errors appeared upon testing. However, the simulated data is
> a bit too basic for me. Would it be possible to rename the column names of
> simulated data frame (eg X1… into dummy risk factors and Y1-Y300 into dummy
> sample/patient names?). Moreover, I dont see whether Y1-Y300 scores are
> actually taken into account in any of the functionalities. If not, I would
> remove these from the simulated dataframe, an otherwise, I would include an
> extra argument that refers to observed sample scores for the functions that
> need sample observations (I guess all?). Ideally, would it be feasible to
> include Cancer and Steroid Hormone (CASH) study and the Womens’
> Contraceptive and Reproductive Experiences (CARE) real you analysed
> previously to analyse “real-world” datasets.
>
> Thanks for your work!
> All the best,
> Arne
>
