### 2019-03-27 e62bb20
revisions in response to JOSS review
