+++ DESCRIPTION
@@ -3,7 +3,7 @@ Type: Package
 Title: Functions to Study Etiologic Heterogeneity
 Version: 0.1.1
 Authors@R: person(given = "Emily C.", family = "Zabor", email = "zabore@mskcc.org", role = c("aut", "cre"))
-Description: A collection of functions related to the study of etiologic heterogeneity both across disease subtypes and across individual disease markers. The included functions allow one to quantify the extent of etiologic heterogeneity in the context of a case-control study, and provide p-values to test for etiologic heterogeneity across individual risk factors.
+Description: A collection of functions related to the study of etiologic heterogeneity both across disease subtypes and across individual disease markers. The included functions allow one to quantify the extent of etiologic heterogeneity in the context of a case-control study, and provide p-values to test for etiologic heterogeneity across individual risk factors. Begg CB, Zabor EC, Bernstein JL, Bernstein L, Press MF, Seshan VE (2013) <doi: 10.1002/sim.5902>.
 Depends:
     R (>= 3.5)
 License: GPL-2
@@ -13,9 +13,7 @@ Encoding: UTF-8
 Imports:
     aod, 
     mlogit, 
-    gtools,
-    knitr,
-    usethis
+    gtools
 Language: en-US
 LazyData: true
 Roxygen: list(markdown = TRUE)
@@ -25,4 +23,6 @@ Suggests:
     testthat,
     covr,
     rmarkdown,
-    dplyr
+    dplyr,
+    knitr,
+    usethis
