+++ _pkgdown.yml
@@ -9,7 +9,7 @@ navbar:
   left:
   - icon: fa-home
     href: index.html
-  - text: Reference
+  - text: Function Documentation
     href: reference/index.html
   - text: Tutorials
     menu:
