### 2019-03-27 #1269 @zabore
Good morning,
I have edited the README to included the main references for the
methodology, and I changed the website heading from "Reference" to
"Function documentation" for clarity. I left the example data variables as
they were, as I am not sure that renaming them adds much. But if I can add
a real world dataset example in the future I will certainly do so.

Thank you,
Emily

On Wed, Mar 27, 2019 at 4:34 AM Arne Van Hoeck <notifications@github.com>
wrote:

> Thanks for your answer. Signed off the authorship as well.
>
> All the best,
> Arne
>
