+++ paper/paper.md
@@ -28,7 +28,7 @@ In previous work we introduced a scalar measure that can be used to quantify the
 
 Separately, we compared available statistical methods for the study of etiologic heterogeneity in case-control studies [@zabor2017]. By unifying the notation of the different methods and employing simulations, we showed that the methods are all able to address two key research questions: 1) whether risk factor effects differ across subtypes of disease and 2) whether risk factor effects differ across levels of each individual disease marker of which the disease subtypes are comprised. Our research also showed that adaptation of polytomous logistic regression has statistical properties at least as good as the more sophisticated methods that have been proposed, provided that the number of disease markers is sufficiently small that the analysis is feasible [@zabor2017]. In R polytomous logistic regression can be implemented using the `multinom` function from the `nnet` package [@venables2002] or the `mlogit` function from the `mlogit` package [@croissant2018], but the additional calculations needed to perform an analysis of etiologic heterogeneity are cumbersome. To facilitate use of this method the R package `riskclustr` introduces functions `eh_test_subtype` and `eh_test_marker` that first fit a standard polytomous logistic regression model using the `mlogit` function from the `mlogit` package [@croissant2018] and then perform additional calculations to address the two preceding questions regarding etiologic heterogeneity. 
 
-Finally, it is increasingly common for statistical or epidemiologic researchers to be confronted with high dimensional disease marker data and when disease subtypes are not pre-defined, this disease marker data can be clustered, and the optimally etiologically heterogeneous subtype solution can be identified. This methodology has been applied to breast cancer [@begg2013; @begg2015] and melanoma [@mauguen2017] and the statistical properties of the approach have been investigated using simulation studies (Zabor et al, under review). In R [@rct2018] unsupervised $k$-means clustering can be implemented using the `kmeans` function in the `stats` package [@rct2018], and the R package`riskclustr` includes a wrapper for the `kmeans` function that calculates the extent of etiologic heterogeneity using the `d` function for each cluster solution resulting from many random starts of the clustering algorithm, and returns the subtype solution that maximizes etiologic heterogeneity. 
+Finally, it is increasingly common for statistical or epidemiologic researchers to be confronted with high dimensional disease marker data and when disease subtypes are not pre-defined, this disease marker data can be clustered, and the optimally etiologically heterogeneous subtype solution can be identified. This methodology has been applied to breast cancer [@begg2013; @begg2015] and melanoma [@mauguen2017] and the statistical properties of the approach have been investigated using simulation studies (Zabor et al, under review). In R [@rct2018] unsupervised $k$-means clustering can be implemented using the `kmeans` function in the `stats` package [@rct2018], and the R package `riskclustr` includes a wrapper for the `kmeans` function that calculates the extent of etiologic heterogeneity using the `d` function for each cluster solution resulting from many random starts of the clustering algorithm, and returns the subtype solution that maximizes etiologic heterogeneity. 
 
 The R package `riskclustr` was designed for use by researchers in epidemiology and biostatistics, and the open-source software package includes  user-friendly tutorials that include examples of how to use the various functions and cover details of all underlying statistical calculations.
 
