### 2019-03-27 #1269 @zabore
I checked the proofs and noticed one typo, just a missing space between two
words, and I pushed a change to correct this in the paper.

I confirm that all references in the bibliography that have an available
DOI have it listed.

I have created a GitHub release and deposited it on Zenodo. DOI:
10.5281/zenodo.2612856

Please let me know if you need anything else from me.

Thank you!
Emily

On Wed, Mar 27, 2019 at 2:43 PM Pjotr Prins <notifications@github.com>
wrote:

> @zabore <https://github.com/zabore> to finalize your submission and
> accept your paper in JOSS, we need two things. First, can you confirm that
> all references in your bibliography have a DOI in the bibliography (if one
> exists).
>
> Second, we need you to deposit a copy of your software repository
> (including any revisions made during the JOSS review process) with a
> data-archiving service.
>
> To do so:
>
>    1. Create a GitHub release
>    <https://help.github.com/articles/creating-releases/> of the current
>    version of your software repository
>    2. Deposit that release with Zenodo <https://zenodo.org/>, figshare
>    <https://figshare.com/>, or a similar DOI issuer.
>    3. Post a comment here with the DOI for the release.
>
