### 2019-03-28 #1269 @danielskatz
👋 @zabore - something in your change causes the paper to no longer compile.

Please try to fix it - once you have made a change, you can enter ```@whedon generate pdf``` as a new comment here.  Once you are happy that the paper compiles and is correct, please let me know, and I'll continue the acceptance process.
