### 2019-03-28 #1269 @labarba
This XML deposit has a _lot_ of false citations, coming from the `.bib` file, and which are not cited in the paper. I've observed several instances of this lately. I don't know how serious it is.
