### 2019-03-28 #1269 @danielskatz
I don't understand this - I thought the PDF generation process would only add to the references when there was a citation in the paper, like bibtex normally does
