### 2019-03-28 #1269 @labarba
@arfon — Please clarify this issue: when authors have many entries in their `.bib` file, they get picked up by `whedon` and added to the XML, which presumably enters false citations for those works in Crossref.
