### 2019-03-28 #1269 @labarba
@danielskatz — I always visually check the XML created after `@whedon accept` and look at the list of citations, then inspect the PDF and count the citations. Sometimes they don't match, and I go back to the authors and ask them to edit the `.bib` file.
