### 2019-03-28 #1269 @labarba
This XML file has 116 citations entries ...
https://github.com/openjournals/joss-papers/blob/bea64c63597ea5a91ad64dc6e622e6eada7c7bcd/joss.01269/10.21105.joss.01269.crossref.xml#L425
