### 2019-03-28 #1269 @danielskatz
It seems a bit awkward, as most authors probably have a .bib file with lots of entries that they don't expect to all show up in the metadata.  Maybe this indicates a bug in whedon?
