### 2019-03-28 #1269 @arfon
> It seems a bit awkward, as most authors probably have a .bib file with lots of entries that they don't expect to all show up in the metadata. Maybe this indicates a bug in whedon?

Let's file this as a bug with Whedon. I don't have an immediate fix unfortunately as we currently take whatever is in the bibtex and generate the Crossref citation metadata based on that. For now, we need to do the check that @labarba is suggesting sorry.
