### 2019-03-28 #1269 @danielskatz
👋 @zabore - Can you remove the entries from your .bib file that are not actually being cited in the paper?  Then we will regenerate the backend XML - This should not change your paper or anything visible to a reader/user.
