### 2019-03-29 #1269 @arfon
> 👋 @zabore - Can you remove the entries from your .bib file that are not actually being cited in the paper? Then we will regenerate the backend XML - This should not change your paper or anything visible to a reader/user.

@zabore - please see the output from @whedon above listing the extraneous references in your `paper.bib`
