### 2019-03-29 #1269 @zabore
Yes I do typically work from a larger master .bib file but I will remove
all unused references and let you know when it's complete.
