### 2019-03-31 #1269 @pjotrp
Thank you @zabore and @avhoeck for your contribution to JOSS!
